Changed int to size_t

+++ homomorphism/src/graph.cpp
@@ -15,7 +15,7 @@ bool has_suffix(const std::string& str, const std::string& suffix)
         str.compare(str.size() - suffix.size(), suffix.size(), suffix) == 0;
 }
 
-void Graph::addEdge(int u, int v) {
+void Graph::addEdge(size_t u, size_t v) {
     if (!matrix_[u * vertices_ + v]) edges_++;
     matrix_[u * vertices_ + v] = true;
     matrix_[v * vertices_ + u] = true;
@@ -27,11 +27,11 @@ std::shared_ptr<Graph> parseGr(std::ifstream& input) {
         if(!std::getline(input, line)) return nullptr;
     } while (line[0] == 'c');
     
-    int n, m;
+    size_t n, m;
     if (!std::sscanf(line.c_str(), "p tw %d %d", &n, &m)) return nullptr;
     std::shared_ptr <Graph> G = std::make_shared<Graph>(n);
 
-    int u, v;
+    size_t u, v;
     while (getline(input, line)) {
         if (line.empty() || line[0] == 'c') continue;
 
@@ -64,12 +64,12 @@ std::shared_ptr<Graph> Graph::fromFile(std::string path) {
 
 std::shared_ptr<Graph> Graph::testGraph()
 {
-    int v = 4;
-    int e = 5;
+    size_t v = 4;
+    size_t e = 5;
     return std::make_shared<Graph>(v, e);
 }
 
-bool Graph::edgeExist(int u, int v)
+bool Graph::edgeExist(size_t u, size_t v)
 {
     return matrix_[u * vertices_ + v];
 }
@@ -80,12 +80,12 @@ bool Graph::isIsomorphic(std::shared_ptr<Graph> g)
     return false;
 }
 
-int Graph::vertCount()
+size_t Graph::vertCount()
 {
     return vertices_;
 }
 
-int Graph::edgeCount()
+size_t Graph::edgeCount()
 {
     return edges_;
 }
@@ -103,17 +103,17 @@ std::string Graph::toGraph6()
     return "";
 }
 
-std::shared_ptr<Graph> Graph::partition(std::set<int>* parts, int size) {
+std::shared_ptr<Graph> Graph::partition(std::set<size_t>* parts, size_t size) {
     std::shared_ptr<Graph> newGraph = std::make_shared<Graph>(size);
 
     bool edgeFound;
-    std::set<int>::iterator uit, uend, vit, vend;
+    std::set<size_t>::iterator uit, uend, vit, vend;
 
-    for(int u = 0; u < size; u++)
+    for(size_t u = 0; u < size; u++)
     {
         uend = parts[u].end();
         
-        for (int v = u + 1; v < size; v++)
+        for (size_t v = u + 1; v < size; v++)
         {
             edgeFound = false;
             uit = parts[u].begin();
@@ -144,9 +144,9 @@ std::shared_ptr<Graph> Graph::partition(std::set<int>* parts, int size) {
 void Graph::prettyPrint(std::ostream& os) 
 {
     os << "Graph with " << vertices_ << " vertices and " << edges_ << " edges:\n";
-    for (int u = 0; u < vertices_; u++) {
+    for (size_t u = 0; u < vertices_; u++) {
         os << "[ ";
-        for (int v = 0; v < vertices_; v++) {
+        for (size_t v = 0; v < vertices_; v++) {
             os << edgeExist(u, v) << " ";
         }
         os << "]" << std::endl;
